+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/store/common/SubpartitionSegmentIndexTracker.java
@@ -37,7 +37,7 @@ public class SubpartitionSegmentIndexTracker {
     // amount of data exceeds the threshold, the segment is switched. Different subpartitions
     // may have duplicate segment indexes, so it is necessary to distinguish different
     // subpartitions when determining whether a tier contains the segment data.
-    private final HashMap<Integer, HashSet<Integer>> subpartitionSegmentIndexes;
+    private final HashMap<Integer, HashSet<Long>> subpartitionSegmentIndexes;
 
     private final Lock[] locks;
 
@@ -55,7 +55,7 @@ public class SubpartitionSegmentIndexTracker {
     }
 
     // Return true if this segment tracker did not already contain the specified segment index.
-    public boolean addSubpartitionSegmentIndex(int subpartitionId, int segmentIndex) {
+    public boolean addSubpartitionSegmentIndex(int subpartitionId, long segmentIndex) {
         if (isBroadCastOnly) {
             return callWithSubpartitionLock(
                     0,
@@ -67,12 +67,12 @@ public class SubpartitionSegmentIndexTracker {
         }
     }
 
-    public boolean hasCurrentSegment(int subpartitionId, int segmentIndex) {
+    public boolean hasCurrentSegment(int subpartitionId, long segmentIndex) {
         if (isBroadCastOnly) {
             return callWithSubpartitionLock(
                     0,
                     () -> {
-                        Set<Integer> segmentIndexes = subpartitionSegmentIndexes.get(0);
+                        Set<Long> segmentIndexes = subpartitionSegmentIndexes.get(0);
                         if (segmentIndexes == null) {
                             return false;
                         }
@@ -82,7 +82,7 @@ public class SubpartitionSegmentIndexTracker {
             return callWithSubpartitionLock(
                     subpartitionId,
                     () -> {
-                        Set<Integer> segmentIndexes;
+                        Set<Long> segmentIndexes;
                         segmentIndexes = subpartitionSegmentIndexes.get(subpartitionId);
                         if (segmentIndexes == null) {
                             return false;
