+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/store/tier/remote/RemoteDataManager.java
@@ -63,7 +63,7 @@ public class RemoteDataManager implements SingleTierDataGate {
     public void setNumBytesInASegment(int numBytesInASegment) {}
 
     @Override
-    public boolean hasCurrentSegment(int subpartitionId, int segmentIndex) {
+    public boolean hasCurrentSegment(int subpartitionId, long segmentIndex) {
         return false;
     }
